Fix more icc remarks
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@119044 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: e7fb8ff1da36c6a4a62482800b39a80b4b20b8bb

+++ Rec/Rich/RichIFBAnalysis/src/HPDIonFeedbackMoni.cpp
@@ -207,12 +207,12 @@ StatusCode HPDIonFeedbackMoni::execute()
   } // ----- End of Loop over L1 boards
 
   if(r1clustersThisEvent>0) {      // If clusters were found in RICH1
-    richHisto1D( HID("R1ClusOcc") ) -> fill(r1clustersThisEvent);
+    richHisto1D( HID("R1ClusOcc") ) -> fill((double)r1clustersThisEvent);
     r1clustersThisEvent = 0;
   }
 
   if(r2clustersThisEvent>0) {      // If clusters were found in RICH2
-    richHisto1D( HID("R2ClusOcc") ) -> fill(r2clustersThisEvent);
+    richHisto1D( HID("R2ClusOcc") ) -> fill((double)r2clustersThisEvent);
     r2clustersThisEvent = 0;
   }
